### 2018-08-25 52fffa2
correcting bib entries and title
