+++ docs/paper.bib
@@ -102,7 +102,12 @@
   year={2014},
   publisher={Elsevier}
 }
-
+@article{brett2016nibabel,
+  title={nibabel: Access a cacophony of neuro-imaging file formats, version 2.1. 0},
+  author={Brett, Matthew and Hanke, Michael and Cipollini, Ben and C{\^o}t{\'e}, Marc-Alexandre and Markiewicz, Chris and Gerhard, Stephan and Larson, Eric and Lee, Gregory R and Halchenko, Yaroslav and Kastman, Erik and others},
+  journal={Zenodo},
+  year={2016}
+}
 @article {Raamana2017biorxiv,
 	author = {Raamana, Pradeep Reddy and Strother, Stephen C.},
 	title = {Impact of spatial scale and edge weight on predictive power of cortical thickness networks},
