### 2018-08-25 eabe69c
results howto and improved docs
