+++ docs/modules.rst
@@ -1,7 +1,0 @@
-graynet
-=======
-
-.. toctree::
-   :maxdepth: 4
-
-   graynet
