+++ docs/getstarted.rst
@@ -2,15 +2,18 @@
 Getting started
 ---------------
 
-``graynet`` currently offer the following streams of processing:
+``graynet`` currently offers the following streams of processing:
 
 
-  - :doc:`cortical` : using vertex-wise ROIs defined on the cortex. This is useful to analyze network-level features based on cortical thickness, curvature, sulcal depth and gyrification. Base-level features can easily obtained from running `Freesurfer <https://surfer.nmr.mgh.harvard.edu/>`_, `CIVET <http://www.bic.mni.mcgill.ca/ServicesSoftware/BasicUsageOfCIVET>`_ or `ANTs <http://stnava.github.io/ANTs/>`_.
+  - :doc:`cortical` : using vertex-wise ROIs defined on the cortex. This is useful to analyze network-level features based on cortical thickness, curvature, sulcal depth and gyrification. Base-level features can easily obtained from running `Freesurfer <https://surfer.nmr.mgh.harvard.edu/>`_
 
-We plan to offer the following streams of processing soon:
+We plan to offer the following soon:
 
 
-  - :doc:`volumetric` : using voxel-wise ROIs defined over the whole brain relying on a volumetric atlas. This is useful to analyze network-level features based on gray matter density dervied from voxel-based morphometry (VBM) or similar approaches. Base-level features can easily obtained from running `SPM <http://www.fil.ion.ucl.ac.uk/spm/software/spm12/>`_, `ANTs <http://stnava.github.io/ANTs/>`_, `FSL <https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FSLVBM>`_, `Freesurfer <https://surfer.nmr.mgh.harvard.edu/>`_, or another suitable package.
+  - :doc:`volumetric` stream : using voxel-wise ROIs defined over the whole brain relying on a volumetric atlas. This is useful to analyze network-level features based on gray matter density dervied from voxel-based morphometry (VBM) or similar approaches. Base-level features can easily obtained from running `SPM <http://www.fil.ion.ucl.ac.uk/spm/software/spm12/>`_, `ANTs <http://stnava.github.io/ANTs/>`_, `FSL <https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FSLVBM>`_, `Freesurfer <https://surfer.nmr.mgh.harvard.edu/>`_, or another suitable package.
+  - Support for additional input formats for the :doc:`cortical` stream:
+    - `CIVET <http://www.bic.mni.mcgill.ca/ServicesSoftware/BasicUsageOfCIVET>`_
+    - `ANTs <http://stnava.github.io/ANTs/>`_.
 
 In both streams, in addition to the computation of pair-wise network-level features, ``graynet`` will help you compute ROI-wise statistics (individual, not pair-wise) for visualization (median thickness in PCG, or variance in GM density within amygdala), as well as to serve as a baseline for network-level features.
 
