+++ docs/installation.rst
@@ -22,3 +22,14 @@ Requirements
  - numpy
  - nibabel
  - hiwenet
+
+
+Citation
+--------
+
+If you found any parts of graynet to be useful in your research, I'd appreciate if you could cite one or more of the following papers in that order:
+
+ - Raamana, P. R. & Strother, S. C., (2018). *graynet: single-subject morphometric networks for neuroscience connectivity applications* (Version 0.3.9). Zenodo. http://doi.org/10.5281/zenodo.1403304
+ - Pradeep Reddy Raamana. *graynet: Subject-wise networks from structural shape and morphological features*. Zenodo. http://doi.org/10.5281/zenodo.997358
+ - Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381.
+ - Raamana, P. R., Weiner, M. W., Wang, L., Beg, M. F., & Alzheimer's Disease Neuroimaging Initiative. (2015). Thickness network features for prognostic applications in dementia. Neurobiology of aging, 36, S91-S102.
