### 2018-08-25 fa6fdb5
adding generated docs to top level
