+++ _sources/about.rst.txt
@@ -1,0 +1,37 @@
+--------------------------------------------------------------------------------------------------
+About
+--------------------------------------------------------------------------------------------------
+
+``graynet`` helps you easily extract single-subject network-level features from T1-weighted mri features such as cortical thickness, gray matter density, subcortical morphometric features, gyrification and curvature.
+
+
+.. image:: https://travis-ci.org/raamana/graynet.svg?branch=master
+   :target: https://travis-ci.org/raamana/graynet
+.. image:: https://zenodo.org/badge/98567518.svg
+   :target: https://zenodo.org/badge/latestdoi/98567518
+.. image:: https://landscape.io/github/raamana/graynet/master/landscape.svg?style=flat
+   :target: https://landscape.io/github/raamana/graynet/master
+   :alt: Code Health
+.. image:: https://api.codacy.com/project/badge/Grade/cffd80f290544e2e824011bfccf35ff8
+   :target: https://www.codacy.com/app/raamana/graynet?utm_source=github.com&amp;utm_medium=referral&amp;utm_content=raamana/graynet&amp;utm_campaign=Badge_Grade
+.. image:: https://img.shields.io/badge/python-2.7%2C%203.5%2C%203.6-blue.svg
+   :alt: Python versions
+.. image:: https://badge.fury.io/py/graynet.svg
+   :target: https://badge.fury.io/py/graynet
+
+
+Applicability
+-------------
+
+This package can aid in understanding brain function (relationship of GM network-level features with behaviour) and in imaging biomarker development for the detection of many brain disorders such as Alzheimer's, Parkinson's, Schizophrenia and the like. The following publications show typical use cases:
+
+ * Raamana, P.R. and Strother, S.C., 2017, Impact of spatial scale and edge weight on predictive power of cortical thickness networks bioRxiv 170381 http://www.biorxiv.org/content/early/2017/07/31/170381. doi: https://doi.org/10.1101/170381
+ * Tijms, B. M., Seriès, P., Willshaw, D. J., & Lawrie, S. M. (2012). Similarity-based extraction of individual networks from gray matter MRI scans. Cerebral Cortex, 22(7), 1530-1541.
+ * Palaniyappan, L., Park, B., Balain, V., Dangi, R., & Liddle, P. (2015). Abnormalities in structural covariance of cortical gyrification in schizophrenia. Brain Structure and Function, 220(4), 2059-2071.
+ * Xu, J., Zhang, J., Zhang, J., Wang, Y., Zhang, Y., Wang, J., ... & Zhang, Y. (2017). Abnormalities in structural covariance of cortical gyrification in Parkinson's disease. Frontiers in Neuroanatomy, 11.
+ * Raamana, P. R., Wen, W., Kochan, N. A., Brodaty, H., Sachdev, P. S., Wang, L., & Beg, M. F. (2014). Novel ThickNet features for the discrimination of amnestic MCI subtypes. NeuroImage: Clinical, 6, 284-295.
+
+
+Check the :doc:`usage_cli` and :doc:`API` pages, and let me know your comments.
+
+Thanks for checking out. Your feedback will be appreciated.
