+++ _sources/volumetric.rst.txt
@@ -1,0 +1,9 @@
+Volumetric graynet
+------------------
+
+Under development, sorry.
+
+Check back soon, thanks!
+
+
+When available, ``graynet`` will help compute ROI-wise statistics (individual, not pair-wise) for visualization (e.g. mean and variance of GM density within amygdala), as well as to serve as a baseline for network-level features. Use the ``-r`` or ``--roi_stats`` flag to achieve it. Only one type of processing (ROI stats, or network-level features) can be done at a time.
