+++ _sources/API.rst.txt
@@ -1,0 +1,20 @@
+--------------
+API
+--------------
+
+A tutorial-like presentation is available at :doc:`usage_api`, using the following API:
+
+
+API Reference
+--------------
+
+.. automodule:: run_workflow
+   :members: extract, roiwise_stats_indiv
+   :undoc-members:
+   :show-inheritance:
+
+
+.. automodule:: multi_edge
+   :members: extract_multiedge
+   :undoc-members:
+   :show-inheritance:

[…]
