### 2018-08-31 #915 @rougier
By the way, @GaelVaroquaux can be also a potential reviewer.
