### 2018-08-31 #915 @arokem
👋 @GaelVaroquaux: would you be available to review this JOSS paper?
