+++ docs/conf.py
@@ -25,6 +25,7 @@ sys.path.insert(0, os.path.abspath('../../graynet'))
 sys.path.insert(0, os.path.abspath('graynet'))
 
 import sphinx_rtd_theme
+import graynet
 
 # -- General configuration ------------------------------------------------
 
@@ -69,9 +70,9 @@ author = 'Pradeep Reddy Raamana'
 # built documents.
 #
 # The short X.Y version.
-version = '0.3'
+version = graynet.__version__
 # The full version, including alpha/beta/rc tags.
-release = '0.3'
+release = graynet.__version__
 
 # The language for content autogenerated by Sphinx. Refer to documentation
 # for a list of supported languages.
