### 2018-08-31 #915 @GaelVaroquaux
> 👋 @GaelVaroquaux: would you be available to review this JOSS paper?

Sorry, I am way behind everything (including reviews). I've been
traveling for a month straight. It's exhausting!
