### 2018-09-02 #924 @whedon
[REVIEW]: graynet: single-subject morphometric networks for neuroscience connectivity applications
**Submitting author:** @raamana (<a href="http://orcid.org/0000-0003-4662-0558">Pradeep Reddy Raamana</a>)
**Repository:** <a href="https://github.com/raamana/graynet" target ="_blank">https://github.com/raamana/graynet</a>
**Version:** v0.3.9

Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue in the <a href="github.com/raamana/graynet" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers) in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice versa will create corresponding breadcrumb trails in the link target.)

- [x] **Repository:** Is the source code for this software available at the <a target="_blank" href="github.com/raamana/graynet">repository url</a>?
