### 2018-09-02 #915 @arokem
@rougier : please head over to #924 for the reviewer check-list. I'm closing this one.
