### 2018-09-08 #924 @rougier
@raamana I've started the review and I've a few questions:

* I'm checking the docs but it seems there is a problem (I can read it at https://raamana.github.io/graynet/ but it's like the css style is broken. I'm using OSX/firefox to read it.
* Can you add all the missing DOI in the paper?
* In the test script, some tests are commented, is that on purpose?
* How do I check I get the right result from those tests?
