### 2018-09-08 65030c9
css improvement - CLI table wrap
