+++ docs/conf.py
@@ -113,11 +113,11 @@ html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 # so a file named "default.css" will overwrite the builtin "default.css".
 
 html_static_path = ['_static']
-# html_context = {
-#     'css_files': [
-#         '_static/theme_overrides.css',  # override wide tables in RTD theme
-#         ],
-#      }
+html_context = {
+    'css_files': [
+        '_static/theme_overrides.css',  # override wide tables in RTD theme
+        ],
+     }
 
 # -- Options for HTMLHelp output ------------------------------------------
 
