### 2018-09-08 #924 @raamana
Thanks for getting started.

1. docs seem to look fine on my laptop, but are somehow broken on the website. This is a bit beyond my expertise, will followup with Sphinx etc to try why it isn't working. Are you able to follow the text without the neat layout? And confirm necessary info is there? For the time being, you could perhaps open the `index.html` at the top-level in `gh-pages` branch of the repo. Sorry for the inconvenience, but I'll figure out whats going on.
2. paper.md does have DOI for this tool noted: 10.5281/zenodo.1403304 - do you mean DOIs for references?
3. They are meant to help me with executing a specific test in the debugging mode. On the CI server, all tests get executed.
4. to check if the code is passing all the tests, do this:

```python
cd <<path to cloned graynet>>
cd graynet/tests
pytest
```

that will run all tests in `test_graynet.py` and it should report no errors (like `12 passed` below):

```bash
$ 08:38:39 SQuark-3 tests >>  pwd
/Users/Reddy/dev/graynet/graynet/tests
$ 08:38:48 SQuark-3 tests >>  pytest
===================================================================================== test session starts ======================================================================================
platform darwin -- Python 3.6.2, pytest-3.4.2, py-1.5.2, pluggy-0.6.0
rootdir: /Users/Reddy/dev/graynet, inifile:
plugins: hypothesis-3.30.1
collected 12 items

test_graynet.py ............                                                                                                                                                             [100%]

======================================================================================= warnings summary =======================================================================================
graynet/tests/test_graynet.py::test_run_roi_stats_via_API
  /Users/Reddy/anaconda/envs/py36/lib/python3.6/site-packages/scipy/stats/stats.py:2831: FutureWarning:

  Using a non-tuple sequence for multidimensional indexing is deprecated; use `arr[tuple(seq)]` instead of `arr[seq]`. In the future this will be interpreted as an array index, `arr[np.array(seq)]`, which will result either in an error or a different result.

-- Docs: http://doc.pytest.org/en/latest/warnings.html
=========================================================================== 12 passed, 1 warnings in 106.92 seconds ============================================================================

```
