### 2018-09-08 #924 @rougier
1. I'll look a the local version but you might want to fix the online version (such that people want to use your soft)
2. Yes, I was talking about references
3. ok
4. I'll try then.
