### 2018-09-08 #924 @rougier
I cannot build the doc because of:
```
 $ make html
Running Sphinx v1.7.9

Extension error:
Could not import extension sphinxarg.ext (exception: No module named 'sphinxarg')
```
Where does this extension comes from?
