### 2018-09-08 #924 @raamana
you don't need to build them - they are already committed to `gh-pages` branch

can you try (cd-ing to top-level graynet dir):
```
git checkout gh-pages
firefox index.html
```

thanks.
