+++ docs/conf.py
@@ -99,8 +99,29 @@ todo_include_todos = False
 # a list of builtin themes.
 #
 
-html_theme = "sphinx_rtd_theme"
-html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+# html_theme = "sphinx_rtd_theme"
+# html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
+
+
+html_theme = 'p-greenblue'
+from PSphinxTheme import utils
+html_theme_path, html_theme, needs_sphinx = utils.set_psphinxtheme(html_theme)
+
+# html_sidebars = {
+#     '**': ['globaltoc.html', 'localtoc.html']
+# }
+# import cakephp_theme
+# html_theme_path = [cakephp_theme.get_html_theme_path()]
+# html_theme = 'cakephp_theme'
+# extensions = ['cakephp_theme']
+# html_context = {
+#     'maintainer': 'Sphinx-themes test',
+#     'project_pretty_name': 'Sphinx Themes',
+#     'projects': {
+#         'CakePHP Book': 'https://book.cakephp.org/',
+#         'Some other project': 'https://example.com/',
+#     }
+# }
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the
