+++ docs/conf.py
@@ -133,12 +133,12 @@ html_theme_path, html_theme, needs_sphinx = utils.set_psphinxtheme(html_theme)
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 
-html_static_path = ['_static']
-html_context = {
-    'css_files': [
-        '_static/theme_overrides.css',  # override wide tables in RTD theme
-        ],
-     }
+# html_static_path = ['_static']
+# html_context = {
+#     'css_files': [
+#         '_static/theme_overrides.css',  # override wide tables in RTD theme
+#         ],
+#      }
 
 # -- Options for HTMLHelp output ------------------------------------------
 
