### 2018-09-08 #924 @rougier
Works thanks (and documentation is much better)!

Test have passed by the way. You might want to add a subsection in the documentation under installation to tell people how to test.
