### 2018-09-08 #924 @raamana
Users are not expected to run the tests themselves - the badge on the repo

[![Build Status](https://travis-ci.org/raamana/graynet.svg?branch=master)](https://travis-ci.org/raamana/graynet)

should give them confidence that its passing all the tests, and releases are not made if the tests fail.

Developers/contributors typically know what to do ;)

This document provides details:
https://github.com/raamana/graynet/blob/master/CONTRIBUTING.md
