### 2018-09-08 #924 @rougier
Do you think it could be possible to add a small toy example alongside the soft ?
With the missing DOI this is the last thing to do I think.
