### 2018-09-08 #924 @rougier
You're right, but some users might want to test on their own machine to know if their installation is ok.
