### 2018-09-08 #924 @rougier
Sorry. Things to do:

1. Add missing DOI from the references in your paper.md (and regenerate it)
2. ~add a small toy example~ Forget about it, I totally missed the script repo...
