### 2018-09-08 7d35d7f
link to contrib guide and adding developer guidelines
