### 2018-09-08 ed76e3f
updating docs
