### 2018-09-08 #924 @raamana
Thanks - so the doc website seems to have fixed itself with a new theme: https://raamana.github.io/graynet/index.html

Per your suggestion, I've added the developer guidelines: https://github.com/raamana/graynet/blob/gh-pages/docs/contributing.rst

This will show up on the docs in a few minutes or so

I will add the DOIs now.
