### 2018-09-08 31070be
adding DOI to references
