+++ docs/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={774--781},
   year={2012},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2012.01.021}
 }
 
 @article{raamana2015thickness,
@@ -17,6 +18,7 @@
   pages={S91--S102},
   year={2015},
   publisher={Elsevier}
+  doi={10.1016/j.neurobiolaging.2014.05.040}
 }
 
 @article{raamana2017hiwenet,
@@ -36,7 +38,8 @@
   number={7},
   pages={1530--1541},
   year={2012},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/cercor/bhr221}
 }
 @article{palaniyappan2015abnormalities,
   title={Abnormalities in structural covariance of cortical gyrification in schizophrenia},
@@ -46,7 +49,8 @@
   number={4},
   pages={2059--2071},
   year={2015},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s00429-014-0772-2}
 }
 @article{xu2017abnormalities,
   title={Abnormalities in structural covariance of cortical gyrification in Parkinson's disease},
@@ -54,14 +58,16 @@
   journal={Frontiers in Neuroanatomy},
   volume={11},
   year={2017},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/fnana.2017.00012}
 }
 
 @article{hagberg2005networkx,
   title={Networkx: Python software for the analysis of networks},
   author={Hagberg, A and Schult, D and Swart, P},
   journal={Mathematical Modeling and Analysis, Los Alamos National Laboratory},
-  year={2005}
+  year={2005},
+  doi={}
 }
 
 @article{evans2013networks,
@@ -71,7 +77,8 @@
   volume={80},
   pages={489--504},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2013.05.054}
 }
 @article{oliphant2007python,
   title={Python for scientific computing},
@@ -80,7 +87,8 @@
   volume={9},
   number={3},
   year={2007},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2007.58}
 }
 @article{walt2011numpy,
   title={The NumPy array: a structure for efficient numerical computation},
@@ -90,7 +98,8 @@
   number={2},
   pages={22--30},
   year={2011},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/MCSE.2011.37}
 }
 
 @article{raamana2014novel,
@@ -100,13 +109,16 @@
   volume={6},
   pages={284--295},
   year={2014},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.nicl.2014.09.005}
 }
 @article{brett2016nibabel,
   title={nibabel: Access a cacophony of neuro-imaging file formats, version 2.1. 0},
   author={Brett, Matthew and Hanke, Michael and Cipollini, Ben and C{\^o}t{\'e}, Marc-Alexandre and Markiewicz, Chris and Gerhard, Stephan and Larson, Eric and Lee, Gregory R and Halchenko, Yaroslav and Kastman, Erik and others},
   journal={Zenodo},
-  year={2016}
+  year={2016},
+  doi={10.5281/zenodo.591597}
+
 }
 @article {Raamana2017biorxiv,
 	author = {Raamana, Pradeep Reddy and Strother, Stephen C.},
