+++ docs/paper.bib
@@ -17,7 +17,7 @@
   volume={36},
   pages={S91--S102},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.neurobiolaging.2014.05.040}
 }
 
