### 2018-09-08 #924 @raamana
The article is filled with DOIs now, except for `networkx` package which I couldn't find DOI for (seems like this old paper never got indexed). Let me know what else needs to be done.
