### 2018-09-10 #924 @rougier
Thanks @raamana. I think your paper is ready for publication (I recommend publication).
@arokem Do we need another review?
