### 2018-09-10 #924 @rougier
Last question @raamana: did you find what is wrong with the online documentation? Did you managed to fix it? I can open an issue on the repo (as a reminder).
