### 2018-09-10 #924 @raamana
Thank you for the speedy review, much appreciate your efforts!

Not sure why it didn’t work before, but the doc website seems to have fixed itself with a new theme:

https://raamana.github.io/graynet/index.html

Take a look
