### 2018-09-10 af35111
adding info on testing
