### 2018-09-10 26010ce
adding info on testing
