### 2018-09-10 #924 @raamana
I have also added more info on how to test graynet locally, as well as many links to how to test python packages. It should appear at the end of developer guidelines shortly (depending on github's CDN responsiveness :)..)
