### 2018-09-16 5f1870c
highlighting additional info on testing
