### 2018-09-16 #924 @raamana
I added additional info on how to test python packages, if that might be of any help. Let me know what else needs to be done.
