### 2018-09-16 #924 @arokem
@rougier : is there anything more to do here? I see that the one item remaining on your list is the examples. Was that resolved to your satisfaction?
