### 2018-09-18 f394fd2
adding info on scripts folder
