+++ docs/index.rst
@@ -30,6 +30,7 @@ graynet
    cortical
    volumetric
    usage_cli
+   extra_scripts
    API
    usage_api
    results_howto
