### 2018-09-23 #924 @arokem
@rougier: thank you so much for the review!

@raamana: I believe this paper is ready to be accepted! Could you please create another release of the software and an archive (e.g., on zenodo) and then post the DOI here?
