### 2018-09-23 #924 @raamana
Hi Ariel, there is some issue with Zenodo - its failing to create a release
https://zenodo.org/account/settings/github/repository/raamana/graynet

Don't want to make a distinct new upload, outside the typical mechanism for Zenodo, that might separate this DOI from the previous releases.. How do you handle this?

I can try again in few days.
