### 2018-10-02 #924 @raamana
I tired it again, just now. Luckily it worked. Final DOI is: 10.5281/zenodo.1442846
